Scanner function order changed, removed some redundancy
Just to clean up the code a bit.

In switch statement, instead of break at the end of each case. Changed it to "return 0"  since it already found the type. There's no need to continue further.
Moved the shorter if statement sequences first in the scanner function to save time in case it is one of the short ones. Well mainly just period. Also removed the reserved check boolean variable since it shouldn't be needed if the function returns immediately after it finds out it is a reserved word.

Removed this code comment bit to make the code look more visually appealing and it is redundant since all of the directions are followed anyways
  /* **
  1. If it is eofm, return right now.

  2. Call the token functions (word and period) 
     one after another (if-then-else).
     Generate a lexical error message if both DFAs failed.
     Let the tokentype be ERROR in that case.

  3. If it was a word,
     check against the reservedwords list.
     If not reserved, tokentype is WORD1 or WORD2
     decided based on the last character.

  4. Return the token type & string  (pass by reference)
  */